schema updated

+++ db/migrate/20241125064700_create_contributions.rb
@@ -1,7 +1,7 @@
 class CreateContributions < ActiveRecord::Migration[7.1]
   def change
     create_table :contributions do |t|
-      t.float :amount
+      t.integer :amount
       t.date :date
       t.boolean :paid
       t.references :investment, null: false, foreign_key: true
